Add tests
Provide tests to do some basic checks on the formula. The 'run_tests.sh'
is a bit of a kludge to run the formula multiple times with different
pillar data. It may be worth looking at if this can be implemented in a
better way in the future... but it will do for now.