Manage the system mail aliases
While technically not part of Exim, the system aliases file is used by
Exim. We can probably assume that if the exim-formula is being used,
then the user probably wants to have a way to manage the aliases as
well. We can do this by providing a way to manage aliases using pillar
data.

If the user has another preferred method for managing the aliases then
they can ignore the exim.aliases state.