Manage the /etc/mailname file
The /etc/mailname file is not part of Exim, but it is used by Exim on
Debian based hosts. We can provide an easy way to manage the file using
pillar data if the user chooses to manage the file using the
exim-formula. If the user has another preferred method for managing the
/etc/mailname file they can ignore the exim.mailname state.