Manage conf.d files using pillar
For configurations that require something more that the very basic Exim
setup user will usually need to edit the files in the conf.d (when using
split configuration files). Using pillar we can probably accommodate 90%
of the users needs.

In the future we may need to look at providing support for managing
files directly, but we will wait an see if a use-case for this arises.