confd pillar data contains just the file contents
The previous method of managing the conf.d files was a bit
over-engineered and not very flexible. Storing the contents of conf.d
files in the pillar data gives the users a lot more freedom with how to
manage the contents of these files without having to worry about how
edge cases should be managed.